# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fleet Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /maintenance:
    get:
      operationId: listMaintenance
      tags: [maintenance]
      summary: List all maintenance records
      responses:
        "200":
          description: List of maintenance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Maintenance"
    post:
      operationId: createMaintenance
      tags: [maintenance]
      summary: Create maintenance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaintenanceRequest"
      responses:
        "201":
          description: Maintenance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maintenance"
  /maintenance/{id}:
    get:
      operationId: getMaintenance
      tags: [maintenance]
      summary: Get maintenance by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Maintenance details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maintenance"
    put:
      operationId: updateMaintenance
      tags: [maintenance]
      summary: Update maintenance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMaintenanceRequest"
      responses:
        "200":
          description: Maintenance updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maintenance"
    delete:
      operationId: deleteMaintenance
      tags: [maintenance]
      summary: Delete maintenance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemRequest"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemRequest"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /inventory/movements:
    post:
      operationId: createInventoryMovement
      tags: [inventory]
      summary: Create inventory movement (in/out/adjustment)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryMovementRequest"
      responses:
        "201":
          description: Movement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryMovement"
    get:
      operationId: listInventoryMovements
      tags: [inventory]
      summary: List inventory movements
      responses:
        "200":
          description: List of movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryMovement"
...
  /reports/maintenance:
    get:
      operationId: getMaintenanceReport
      tags: [reports]
      summary: Get maintenance report data
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: vehicleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Maintenance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceReport"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Maintenance:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehiclePlate:
          type: string
          nullable: true
        type:
          type: string
        description:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed]
        cost:
          type: number
          nullable: true
        mileage:
          type: number
          nullable: true
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        partsUsed:
          type: array
          items:
            $ref: "#/components/schemas/MaintenancePart"
        createdAt:
          type: string
      required: [id, vehicleId, type, description, date, status, createdAt]
    MaintenancePart:
      type: object
      properties:
        inventoryItemId:
          type: integer
        itemName:
          type: string
        quantity:
          type: number
        unitCost:
          type: number
      required: [inventoryItemId, itemName, quantity, unitCost]
    CreateMaintenanceRequest:
      type: object
      properties:
        vehicleId:
          type: integer
        type:
          type: string
        description:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed]
        cost:
          type: number
          nullable: true
        mileage:
          type: number
          nullable: true
        supplierId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        partsUsed:
          type: array
          items:
            $ref: "#/components/schemas/MaintenancePart"
      required: [vehicleId, type, description, date, status]
    UpdateMaintenanceRequest:
      type: object
      properties:
        vehicleId:
          type: integer
        type:
          type: string
        description:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed]
        cost:
          type: number
          nullable: true
        mileage:
          type: number
          nullable: true
        supplierId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        partsUsed:
          type: array
          items:
            $ref: "#/components/schemas/MaintenancePart"
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        minStock:
          type: number
        unitPrice:
          type: number
          nullable: true
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, category, unit, currentStock, minStock, createdAt]
    CreateInventoryItemRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        minStock:
          type: number
        unitPrice:
          type: number
          nullable: true
        supplierId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, category, unit, currentStock, minStock]
    UpdateInventoryItemRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        minStock:
          type: number
        unitPrice:
          type: number
          nullable: true
        supplierId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
    InventoryMovement:
      type: object
      properties:
        id:
          type: integer
        inventoryItemId:
          type: integer
        itemName:
          type: string
          nullable: true
        type:
          type: string
          enum: [in, out, adjustment]
        quantity:
          type: number
        reason:
          type: string
          nullable: true
        maintenanceId:
          type: integer
          nullable: true
        date:
          type: string
        createdAt:
          type: string
      required: [id, inventoryItemId, type, quantity, date, createdAt]
    CreateInventoryMovementRequest:
      type: object
      properties:
        inventoryItemId:
          type: integer
        type:
          type: string
          enum: [in, out, adjustment]
        quantity:
          type: number
        reason:
          type: string
          nullable: true
        maintenanceId:
          type: integer
          nullable: true
        date:
          type: string
      required: [inventoryItemId, type, quantity, date]
...
    MaintenanceReport:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/Maintenance"
        totalCost:
          type: number
        byVehicle:
          type: array
          items:
            type: object
            properties:
              vehiclePlate:
                type: string
              totalCost:
                type: number
              count:
                type: integer
            required: [vehiclePlate, totalCost, count]
        byType:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              count:
                type: integer
              totalCost:
                type: number
            required: [type, count, totalCost]
      required: [records, totalCost, byVehicle, byType]
...