# lib/api-zod/src/generated/types/createInventoryItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fleet Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInventoryItemRequest {
  name: string;
  category: string;
  unit: string;
  currentStock: number;
  minStock: number;
  unitPrice?: number | null;
  supplierId?: number | null;
  notes?: string | null;
}