# lib/api-zod/src/generated/types/maintenancePart.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fleet Management API
 * OpenAPI spec version: 0.1.0
 */

export interface MaintenancePart {
  inventoryItemId: number;
  itemName: string;
  quantity: number;
  unitCost: number;
}